refactor(test): Modify mocks & tests for packed license token metadata

+++ test/foundry/integration/flows/disputes/Disputes.t.sol
@@ -17,7 +17,7 @@ contract Flows_Integration_Disputes is BaseIntegration {
     using Strings for *;
 
     mapping(uint256 tokenId => address ipAccount) internal ipAcct;
-    uint256 internal ncSocialRemixTermsId;
+    uint32 internal ncSocialRemixTermsId;
 
     function setUp() public override {
         super.setUp();
