+++ test/foundry/modules/licensing/LicensingModule.t.sol
@@ -61,7 +61,7 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_attachLicenseTerms_attachOneLicenseToOneIP() public {
-        uint256 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
+        uint32 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
         vm.expectEmit();
         emit ILicensingModule.LicenseTermsAttached(ipOwner1, ipId1, address(pilTemplate), termsId);
         vm.prank(ipOwner1);
@@ -80,7 +80,7 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_attachLicenseTerms_sameLicenseAttachMultipleIP() public {
-        uint256 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
+        uint32 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
 
         vm.expectEmit();
         emit ILicensingModule.LicenseTermsAttached(ipOwner1, ipId1, address(pilTemplate), termsId);
@@ -117,12 +117,12 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_attachLicenseTerms_DifferentLicensesAttachToSameIP() public {
-        uint256 termsId1 = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
+        uint32 termsId1 = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
         vm.expectEmit();
         emit ILicensingModule.LicenseTermsAttached(ipOwner1, ipId1, address(pilTemplate), termsId1);
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId1);
-        (address licenseTemplate1, uint256 licenseTermsId1) = licenseRegistry.getAttachedLicenseTerms(ipId1, 0);
+        (address licenseTemplate1, uint32 licenseTermsId1) = licenseRegistry.getAttachedLicenseTerms(ipId1, 0);
         assertEq(licenseTemplate1, address(pilTemplate));
         assertEq(licenseTermsId1, termsId1);
         assertTrue(licenseRegistry.hasIpAttachedLicenseTerms(ipId1, address(pilTemplate), termsId1));
@@ -134,12 +134,12 @@ contract LicensingModuleTest is BaseTest {
         assertFalse(licenseRegistry.isDerivativeIp(ipId1));
         assertTrue(licenseRegistry.exists(address(pilTemplate), termsId1));
 
-        uint256 termsId2 = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
+        uint32 termsId2 = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
         vm.expectEmit();
         emit ILicensingModule.LicenseTermsAttached(ipOwner1, ipId1, address(pilTemplate), termsId2);
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId2);
-        (address licenseTemplate2, uint256 licenseTermsId2) = licenseRegistry.getAttachedLicenseTerms(ipId1, 1);
+        (address licenseTemplate2, uint32 licenseTermsId2) = licenseRegistry.getAttachedLicenseTerms(ipId1, 1);
         assertEq(licenseTemplate2, address(pilTemplate));
         assertEq(licenseTermsId2, termsId2);
         assertTrue(licenseRegistry.hasIpAttachedLicenseTerms(ipId1, address(pilTemplate), termsId2));
@@ -153,7 +153,7 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_attachLicenseTerms_revert_licenseNotExist() public {
-        uint256 nonExistTermsId = 9999;
+        uint32 nonExistTermsId = 9999;
         assertFalse(licenseRegistry.exists(address(pilTemplate), nonExistTermsId));
 
         vm.expectRevert(
@@ -190,7 +190,7 @@ contract LicensingModuleTest is BaseTest {
             uri: ""
         });
 
-        uint256 termsId = pilTemplate.registerLicenseTerms(terms);
+        uint32 termsId = pilTemplate.registerLicenseTerms(terms);
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
 
@@ -246,14 +246,14 @@ contract LicensingModuleTest is BaseTest {
         });
 
         vm.warp(11 days);
-        uint256 anotherTermsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
+        uint32 anotherTermsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
         vm.expectRevert(abi.encodeWithSelector(Errors.LicenseRegistry__IpExpired.selector, ipId2));
         vm.prank(ipOwner2);
         licensingModule.attachLicenseTerms(ipId2, address(pilTemplate), anotherTermsId);
     }
 
     function test_LicensingModule_attachLicenseTerms_revert_attachSameLicenseToIpTwice() public {
-        uint256 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
+        uint32 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
         vm.expectRevert(
@@ -269,14 +269,14 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_attachLicenseTerms_revert_attachLicenseDifferentTemplate() public {
-        uint256 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
+        uint32 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
 
         MockLicenseTemplate mockLicenseTemplate = new MockLicenseTemplate();
         vm.prank(admin);
         licenseRegistry.registerLicenseTemplate(address(mockLicenseTemplate));
-        uint256 mockTermsId = mockLicenseTemplate.registerLicenseTerms();
+        uint32 mockTermsId = mockLicenseTemplate.registerLicenseTerms();
         vm.expectRevert(
             abi.encodeWithSelector(
                 Errors.LicenseRegistry__UnmatchedLicenseTemplate.selector,
@@ -291,7 +291,7 @@ contract LicensingModuleTest is BaseTest {
 
     function test_LicensingModule_attachLicenseTerms_revert_Unregistered_LicenseTemplate() public {
         MockLicenseTemplate mockLicenseTemplate = new MockLicenseTemplate();
-        uint256 mockTermsId = mockLicenseTemplate.registerLicenseTerms();
+        uint32 mockTermsId = mockLicenseTemplate.registerLicenseTerms();
         vm.expectRevert(
             abi.encodeWithSelector(
                 Errors.LicensingModule__LicenseTermsNotFound.selector,
@@ -304,7 +304,7 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_attachLicenseTerms_revert_NonExists_LicenseTerms() public {
-        uint256 nonExistsTermsId = 9999;
+        uint32 nonExistsTermsId = 9999;
         vm.expectRevert(
             abi.encodeWithSelector(
                 Errors.LicensingModule__LicenseTermsNotFound.selector,
@@ -317,7 +317,7 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_mintLicenseTokens_singleToken() public {
-        uint256 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
+        uint32 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
 
@@ -344,7 +344,7 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_mintLicenseTokens_mintMultipleTokens() public {
-        uint256 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
+        uint32 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
 
@@ -380,7 +380,7 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_mintLicenseTokens_mintMultipleTimes() public {
-        uint256 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
+        uint32 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
 
@@ -439,7 +439,7 @@ contract LicensingModuleTest is BaseTest {
             uri: ""
         });
 
-        uint256 termsId = pilTemplate.registerLicenseTerms(terms);
+        uint32 termsId = pilTemplate.registerLicenseTerms(terms);
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
 
@@ -462,7 +462,7 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_mintLicenseTokens_revert_invalidInputs() public {
-        uint256 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
+        uint32 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
 
@@ -489,7 +489,7 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_mintLicenseTokens_revert_NonIpOwnerMintNotAttachedLicense() public {
-        uint256 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
+        uint32 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
         address receiver = address(0x111);
         vm.expectRevert(
             abi.encodeWithSelector(
@@ -526,7 +526,7 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_mintLicenseTokens_revert_paused() public {
-        uint256 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
+        uint32 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
 
@@ -547,7 +547,7 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_mintLicenseTokens_IpOwnerMintNotAttachedLicense() public {
-        uint256 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
+        uint32 termsId = pilTemplate.registerLicenseTerms(PILFlavors.defaultValuesLicenseTerms());
         address receiver = address(0x111);
         vm.prank(ipOwner1);
         uint256 lcTokenId = licensingModule.mintLicenseTokens({
@@ -570,7 +570,7 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_registerDerivativeWithLicenseTokens_singleParent() public {
-        uint256 termsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
+        uint32 termsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
 
@@ -618,7 +618,7 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_registerDerivativeWithLicenseTokens_revert_pause() public {
-        uint256 termsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
+        uint32 termsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
 
@@ -642,7 +642,7 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_registerDerivativeWithLicenseTokens_twoParents() public {
-        uint256 termsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
+        uint32 termsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
         vm.prank(ipOwner2);
@@ -721,7 +721,7 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_registerDerivativeWithLicenseTokens_revert_parentIsChild() public {
-        uint256 termsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
+        uint32 termsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
         vm.prank(ipOwner1);
 
         uint256 lcTokenId = licensingModule.mintLicenseTokens({
@@ -764,7 +764,7 @@ contract LicensingModuleTest is BaseTest {
             uri: ""
         });
 
-        uint256 termsId = pilTemplate.registerLicenseTerms(terms);
+        uint32 termsId = pilTemplate.registerLicenseTerms(terms);
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
 
@@ -805,10 +805,10 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_registerDerivativeWithLicenseTokens_revert_ParentExpired() public {
-        uint256 termsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
+        uint32 termsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
         PILTerms memory expiredTerms = PILFlavors.nonCommercialSocialRemixing();
         expiredTerms.expiration = 10 days;
-        uint256 expiredTermsId = pilTemplate.registerLicenseTerms(expiredTerms);
+        uint32 expiredTermsId = pilTemplate.registerLicenseTerms(expiredTerms);
 
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
@@ -885,7 +885,7 @@ contract LicensingModuleTest is BaseTest {
         );
         assertEq(anotherLicenseTemplate, address(pilTemplate));
         assertEq(anotherLicenseTermsId, expiredTermsId);
-        uint256[] memory licenseTerms = new uint256[](2);
+        uint32[] memory licenseTerms = new uint32[](2);
         licenseTerms[0] = termsId;
         licenseTerms[1] = expiredTermsId;
 
@@ -912,7 +912,7 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_registerDerivativeWithLicenseTokens_revert_childAlreadyAttachedLicense() public {
-        uint256 termsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
+        uint32 termsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
 
@@ -934,7 +934,7 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_registerDerivativeWithLicenseTokens_revert_AlreadyRegisteredAsDerivative() public {
-        uint256 termsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
+        uint32 termsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
         vm.prank(ipOwner2);
@@ -971,7 +971,7 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_registerDerivativeWithLicenseTokens_revert_notLicensee() public {
-        uint256 termsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
+        uint32 termsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
 
@@ -995,7 +995,7 @@ contract LicensingModuleTest is BaseTest {
     }
 
     function test_LicensingModule_singleTransfer_verifyOk() public {
-        uint256 termsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
+        uint32 termsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
 
@@ -1062,7 +1062,7 @@ contract LicensingModuleTest is BaseTest {
             uri: ""
         });
 
-        uint256 termsId = pilTemplate.registerLicenseTerms(terms);
+        uint32 termsId = pilTemplate.registerLicenseTerms(terms);
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
 
@@ -1112,7 +1112,7 @@ contract LicensingModuleTest is BaseTest {
             uri: ""
         });
 
-        uint256 termsId = pilTemplate.registerLicenseTerms(terms);
+        uint32 termsId = pilTemplate.registerLicenseTerms(terms);
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
 
@@ -1159,7 +1159,7 @@ contract LicensingModuleTest is BaseTest {
             uri: ""
         });
 
-        uint256 termsId = pilTemplate.registerLicenseTerms(terms);
+        uint32 termsId = pilTemplate.registerLicenseTerms(terms);
         vm.prank(ipOwner1);
         licensingModule.attachLicenseTerms(ipId1, address(pilTemplate), termsId);
 
@@ -1209,7 +1209,7 @@ contract LicensingModuleTest is BaseTest {
     function test_LicensingModule_registerDerivative_revert_emptyParentIpIds() public {
         vm.expectRevert(Errors.LicensingModule__NoParentIp.selector);
         vm.prank(ipOwner2);
-        licensingModule.registerDerivative(ipId2, new address[](0), new uint256[](0), address(0), "");
+        licensingModule.registerDerivative(ipId2, new address[](0), new uint32[](0), address(0), "");
     }
 
     function test_LicensingModule_registerDerivative_revert_parentIdsLengthMismatchWithLicenseIds() public {
@@ -1217,12 +1217,12 @@ contract LicensingModuleTest is BaseTest {
         parentIpIds[0] = ipId1;
         vm.expectRevert(abi.encodeWithSelector(Errors.LicensingModule__LicenseTermsLengthMismatch.selector, 1, 0));
         vm.prank(ipOwner2);
-        licensingModule.registerDerivative(ipId2, parentIpIds, new uint256[](0), address(0), "");
+        licensingModule.registerDerivative(ipId2, parentIpIds, new uint32[](0), address(0), "");
     }
 
     function test_LicensingModule_registerDerivative_revert_IncompatibleLicenses() public {
-        uint256 socialRemixTermsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
-        uint256 commUseTermsId = pilTemplate.registerLicenseTerms(
+        uint32 socialRemixTermsId = pilTemplate.registerLicenseTerms(PILFlavors.nonCommercialSocialRemixing());
+        uint32 commUseTermsId = pilTemplate.registerLicenseTerms(
             PILFlavors.commercialUse({
                 mintingFee: 0,
                 currencyToken: address(erc20),
@@ -1239,7 +1239,7 @@ contract LicensingModuleTest is BaseTest {
         parentIpIds[0] = ipId1;
         parentIpIds[1] = ipId2;
 
-        uint256[] memory licenseTermsIds = new uint256[](2);
+        uint32[] memory licenseTermsIds = new uint32[](2);
         licenseTermsIds[0] = socialRemixTermsId;
         licenseTermsIds[1] = commUseTermsId;
